Rename links and add i18n
