+++ src/app/reports/routes.js
@@ -1,10 +1,10 @@
-import FamilyOrganizationView from './families/by_organization/view';
+import FamilyOrganizationView from './snapshots/by_organization/view';
 
 const reports = props => {
   const { app } = props;
   const routes = {
     appRoutes: {
-        'reports/families/byOrganization(/)': 'showFamiliesByOrgamization'
+        'reports/snapshots/byOrganization(/)': 'showFamiliesByOrgamization'
     },
     controller: {
